Extract LinkedBinder & ScopedBinder to own files
